+++ tests/run-regex.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/kedacore/keda/v2/tests/utils"
-
-func main() {
-	utils.RegexTestFiles()
-}