+++ CHANGELOG.md
@@ -64,6 +64,7 @@ Here is an overview of all new **experimental** features:
 
 ### Improvements
 
+- **General**: Added Pre Regex check before building image in e2e test ([#5783](https://github.com/kedacore/keda/issues/5783))
 - **GCP Scalers**: Added custom time horizon in GCP scalers ([#5778](https://github.com/kedacore/keda/issues/5778))
 
 ### Fixes